requirements.txt
PyJWT~=2.4.0
Flask~=2.0.3
SQLAlchemy~=1.4.47
Werkzeug~=2.0.3